reafactor: config into separate package
